Fix experiments ps_noise addition (#44)
Changes to the configuration and main script for estimating causal
effects. The most important changes involve the addition of a new
configuration parameter and updates to the main script to use this new
parameter.

Configuration updates:

*
[`ehr2vec/configs/example_configs/06_estimate_effect_binary.yaml`](diffhunk://#diff-118c10016daadc981c3ff8b0c1b7882c434afe554300b52c48372f9ea1124505R26):
Added a new parameter `ps_noise` with a value of `0.1`.

Main script updates:

*
[`ehr2vec/main/06_estimate_causal_effect.py`](diffhunk://#diff-b414d8712766e39959dc815fe1695e03c6f7f57fd52427f9af0fc8456745a40fL111-R114):
Updated the script to use the new `ps_noise` parameter instead of the
old `ps noise` parameter for adding noise to propensity scores. This
includes changes to the logging and the calculation of propensity
scores.